requirements.txt
requests
requests-toolbelt
lxml
python-dotenv
rrdtool
Flask
Jinja2